requirements.txt
numpy==1.26.4
chromadb==0.4.22
fastembed==0.7.4
fastapi==0.123.9
uvicorn==0.38.0
python-dotenv==1.2.1
groq==0.37.1
PyPDF2==3.0.1
python-multipart==0.0.18
pycryptodome
